Fix footer behaving as child of action-call-card
div element with the "footer" class was behaving as a child element of
the div element with the "action-call-card" class. The reason for this
was because the div with the class "action-call-button", which is a
child of the "action-call-card" class, didn't have a closing tag.
Resulting in everything below being a child of said class, simply added
a closing div tag to fix the issue.